Use typedefs for structs

+++ include/kernel/vga.h
@@ -24,8 +24,8 @@ enum vga_color {
 	VGA_COLOR_WHITE = 15,
 };
 
-static const size_t VGA_WIDTH = 80;
-static const size_t VGA_HEIGHT = 25;
+static const uint8_t VGA_WIDTH = 80;
+static const uint8_t VGA_HEIGHT = 25;
 
 void vga_init();
 void vga_putch(char c);
